HAPPY-5906 some imrovement

+++ dpdocter/src/main/java/com/dpdocter/collections/DoctorSubscriptionPaymentCollection.java
@@ -11,8 +11,8 @@ import com.dpdocter.enums.PackageType;
 import com.dpdocter.enums.PaymentMode;
 
 @Document(collection = "subscription_payment_cl")
-public class DoctorSubscriptionPaymentCollection extends GenericCollection{
-	
+public class DoctorSubscriptionPaymentCollection extends GenericCollection {
+
 	@Field
 	private ObjectId id;
 
@@ -32,10 +32,10 @@ public class DoctorSubscriptionPaymentCollection extends GenericCollection{
 	private Discount discount;
 
 	@Field
-	private Double amount = 0.0;
+	private int amount = 0;
 
 	@Field
-	private Double discountAmount = 0.0;
+	private int discountAmount = 0;
 
 	@Field
 	private Double transferAmount = 0.0;
@@ -60,7 +60,7 @@ public class DoctorSubscriptionPaymentCollection extends GenericCollection{
 
 	@Field
 	private PackageType packageName;
-	
+
 	@Field
 	private String orderId;
 	@Field
@@ -114,19 +114,19 @@ public class DoctorSubscriptionPaymentCollection extends GenericCollection{
 		this.discount = discount;
 	}
 
-	public Double getAmount() {
+	public int getAmount() {
 		return amount;
 	}
 
-	public void setAmount(Double amount) {
+	public void setAmount(int amount) {
 		this.amount = amount;
 	}
 
-	public Double getDiscountAmount() {
+	public int getDiscountAmount() {
 		return discountAmount;
 	}
 
-	public void setDiscountAmount(Double discountAmount) {
+	public void setDiscountAmount(int discountAmount) {
 		this.discountAmount = discountAmount;
 	}
 
@@ -209,7 +209,5 @@ public class DoctorSubscriptionPaymentCollection extends GenericCollection{
 	public void setReciept(String reciept) {
 		this.reciept = reciept;
 	}
-	
-	
 
 }
